+++ fs2es_indexer/Fs2EsIndexer.py
@@ -19,13 +19,14 @@ class Fs2EsIndexer(object):
         self.elasticsearch_url = elasticsearch_config.get('url', 'http://localhost:9200')
         self.elasticsearch_index = elasticsearch_config.get('index', 'files')
         self.elasticsearch_bulk_size = elasticsearch_config.get('bulk_size', 10000)
+        self.elasticsearch_lib_version = elasticsearch_config.get('library_version', 8)
         self.exclusion_strings = exclusions.get('partial_paths', [])
         self.exclusion_reg_exps = exclusions.get('regular_expressions', [])
 
-        # The version of the elasticsearch-library
-        # V8 has some pretty big changes, so we need to switch some statements.
-        # Currently hardcoded to v7 in setup.py
-        self.elasticsearch_version = 7
+        if self.elasticsearch_lib_version != 7 and self.elasticsearch_lib_version != 8:
+            self.print(
+                'This tool only works with the elasticsearch library v7 or v8. Your configured version "%s" is not supported currently.' % self.elasticsearch_lib_version
+            )
 
         if 'user' in elasticsearch_config:
             elasticsearch_auth = (elasticsearch_config['user'], elasticsearch_config['password'])
@@ -33,14 +34,13 @@ class Fs2EsIndexer(object):
             elasticsearch_auth = None
 
         self.elasticsearch = elasticsearch.Elasticsearch(
-            hosts=self.elasticsearch_url,
-            http_auth=elasticsearch_auth,
-            max_retries=10,
-            retry_on_timeout=True,
-            use_ssl=elasticsearch_config.get('use_ssl', False),
-            verify_certs=elasticsearch_config.get('use_ssl', True),
-            ssl_show_warn=elasticsearch_config.get('ssl_show_warn', True),
-            ca_certs=elasticsearch_config.get('use_ssl', None)
+            hosts = self.elasticsearch_url,
+            http_auth = elasticsearch_auth,
+            max_retries = 10,
+            retry_on_timeout = True,
+            verify_certs = elasticsearch_config.get('verify_certs', True),
+            ssl_show_warn = elasticsearch_config.get('ssl_show_warn', True),
+            ca_certs = elasticsearch_config.get('ca_certs', None)
         )
 
     def map_path_to_es_document(self, path, filename, index_time):
@@ -128,13 +128,13 @@ class Fs2EsIndexer(object):
         if self.elasticsearch.indices.exists(index=self.elasticsearch_index):
             try:
                 self.print('- Updating mapping of index "%s" ...' % self.elasticsearch_index)
-                if self.elasticsearch_version == 7:
+                if self.elasticsearch_lib_version == 7:
                     self.elasticsearch.indices.put_mapping(
                         index=self.elasticsearch_index,
                         doc_type=None,
                         body=index_mapping['mappings']
                     )
-                elif self.elasticsearch_version == 8:
+                elif self.elasticsearch_lib_version == 8:
                     self.elasticsearch.indices.put_mapping(
                         index=self.elasticsearch_index,
                         doc_type=None,
@@ -152,12 +152,12 @@ class Fs2EsIndexer(object):
             self.print('- Creating index "%s" ...' % self.elasticsearch_index)
 
             try:
-                if self.elasticsearch_version == 7:
+                if self.elasticsearch_lib_version == 7:
                     self.elasticsearch.indices.create(
                         index=self.elasticsearch_index,
                         body=index_mapping
                     )
-                elif self.elasticsearch_version == 8:
+                elif self.elasticsearch_lib_version == 8:
                     self.elasticsearch.indices.create(
                         index=self.elasticsearch_index,
                         mappings=index_mapping['mappings']
@@ -258,7 +258,7 @@ class Fs2EsIndexer(object):
 
         self.print('- Deleting old documents from "%s" ...' % self.elasticsearch_index)
         try:
-            if self.elasticsearch_version == 7:
+            if self.elasticsearch_lib_version == 7:
                 resp = self.elasticsearch.delete_by_query(
                     index=self.elasticsearch_index,
                     body={
@@ -271,7 +271,7 @@ class Fs2EsIndexer(object):
                         }
                     }
                 )
-            elif self.elasticsearch_version == 8:
+            elif self.elasticsearch_lib_version == 8:
                 resp = self.elasticsearch.delete_by_query(
                     index=self.elasticsearch_index,
                     query={
@@ -298,12 +298,12 @@ class Fs2EsIndexer(object):
         """ Deletes all documents in the elasticsearch index """
         self.print('- Deleting all documents from index "%s" ...' % self.elasticsearch_index)
         try:
-            if self.elasticsearch_version == 7:
+            if self.elasticsearch_lib_version == 7:
                 resp = self.elasticsearch.delete_by_query(
                     index=self.elasticsearch_index,
                     body={"query": {"match_all": {}}}
                 )
-            elif self.elasticsearch_version == 8:
+            elif self.elasticsearch_lib_version == 8:
                 resp = self.elasticsearch.delete_by_query(
                     index=self.elasticsearch_index,
                     query={"match_all": {}}
@@ -368,7 +368,7 @@ class Fs2EsIndexer(object):
             }
 
         try:
-            if self.elasticsearch_version == 7:
+            if self.elasticsearch_lib_version == 7:
                 resp = self.elasticsearch.search(
                     index=self.elasticsearch_index,
                     body={
@@ -377,7 +377,7 @@ class Fs2EsIndexer(object):
                     from_=0,
                     size=100
                 )
-            elif self.elasticsearch_version == 8:
+            elif self.elasticsearch_lib_version == 8:
                 resp = self.elasticsearch.search(
                     index=self.elasticsearch_index,
                     query=query,