+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # FileSystem To Elastic Search Indexer Changelog
 
+## 0.4.0
+- Switch to ES-Lib v8 for ElasticSearch 8.0+
+- Add configuration which library version is currently in use
+- Fix problems in ES-Lib v8
+- Add README.md section on how to enable the user authentication
+- Remove "use_ssl" from the ES-constructor and from the configuration
+
 ## 0.3.5
 - Dont throw an error and abort if a file is deleted during indexing (2nd try)